+++ NEWS.md
@@ -5,6 +5,8 @@
   
 * Added `trim_amount()` for `recipes::step_impute_mean()`.
 
+* Added `validation_set_prop()` for `embed::step_discretize_xgb()`.
+
 # dials 1.1.0
 
 ## New parameters